test: Improve tests for REST API

+++ app/adapters/api/src/main/java/it/mulders/traqqr/api/measurements/dto/MeasurementDto.java
@@ -3,7 +3,7 @@ package it.mulders.traqqr.api.measurements.dto;
 import java.time.OffsetDateTime;
 
 public record MeasurementDto(OffsetDateTime timestamp, int odometer, BatteryDto battery, LocationDto location) {
-    public record BatteryDto(int soc) {}
+    public record BatteryDto(byte soc) {}
 
     public record LocationDto(double lat, double lon) {}
 }
